// src/app/api/model/accountSeller.ts
/**
 * 
 * 
 *
 * 
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */


export interface AccountSeller {
    id?: number;
    name?: string;
    lastname?: string;
    email?: string;
    documentNumber?: string;
    phoneNumber?: string;
    documentId?: number;
    documentName?: string;
    zipCodeId?: number;
    zipCodeValue?: string;
    provinceId?: number;
    provinceName?: string;
    locationId?: number;
    locationName?: string;
    businessNameId?: number;
    businessName?: string;
    otherBusinessName?: string;
    addressId?: number;
    addressName?: string;
    otherAddress?: string;
    coupons?: Array<string>;
    isVerified?: boolean;
    countCoupons?: number;
    userId?: number;
    createDate?: any;
}