// src/app/api/model/accountSellerFilter.ts
/**
 * 
 * 
 *
 * 
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */


export interface AccountSellerFilter {
    maxResults?: number;
    firstResult?: number;
    total?: number;
    likeNameBusinessName?: string;
    provinceId?: number;
    locationId?: number;
    businessNameId?: number;
    likeBuyer?: string;
    likeSeller?: string;
    orderBy?: string;
}